requirements.txt
numpy==1.24.4
Pillow==10.1.0
timm==0.9.10
torch==2.4.0
torchvision==0.19.0
transformers==4.37.2
sentencepiece==0.1.99
accelerate==0.30.1
peft==0.12.0

#packaging==23.2
#addict==2.4.0
#editdistance==0.6.2
#einops==0.7.0
#fairscale==0.4.0
#jsonlines==4.0.0
#markdown2==2.4.10
#matplotlib==3.7.4
#more_itertools==10.1.0
#nltk==3.8.1
#opencv_python_headless==4.5.5.64
#openpyxl==3.1.2
#sacrebleu==2.3.2
#seaborn==0.13.0
#shortuuid==1.0.11
#tqdm==4.66.1
#protobuf==4.25.0
#typing_extensions==4.8.0
#uvicorn==0.24.0.post1
#socksio==1.0.0
#gradio==4.41.0
#gradio_client
#decord

#xformers==0.0.22.post7
#flash_attn==2.3.4